+++ CHANGELOG.md
@@ -3,10 +3,10 @@
 ## 0.10.0 (unreleased)
 
 - BREAKING:
-  - Rename `shadow_attrs` argument to `forward_attrs`
-  - Enforce that the same object is not added multiple times to one parent
+  - Rename `shadow_attrs` argument to `forward_attrs`.
+  - Enforce that the same object is not added multiple times to one parent.
   - Rename `GenericNodeData` to `DictWrapper` and remove support for attribut access.
-- tree.to_rdf() is now available for Tree (not only TypedTree)
+- tree.to_rdf() is now available for Tree (not only TypedTree).
 - New method `node.up()` allows method chaining when adding nodes.
 - Passes more pyright 'basic' checks.