+++ CHANGELOG.md
@@ -6,6 +6,7 @@
   - Rename `shadow_attrs` argument to `forward_attrs`.
   - Enforce that the same object is not added multiple times to one parent.
   - Rename `GenericNodeData` to `DictWrapper` and remove support for attribut access.
+  - Drop support for Python 3.8
 - tree.to_rdf() is now available for Tree (not only TypedTree).
 - New method `node.up()` allows method chaining when adding nodes.
 - Passes more pyright 'basic' checks.